requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.4
django-browser-reload==1.17.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-restframework==0.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
kombu==5.4.2
Markdown==3.7
pillow==11.0.0
prompt_toolkit==3.0.48
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
